refactor: `src/shared/`
